refactor(app): Moved components out of apps to components folder
