+++ src/app/business/page.tsx
@@ -4,13 +4,13 @@ import { CheckCircle2, ChevronRight } from "lucide-react";
 import Link from "next/link";
 import Navbar from "@/components/business/Navbar";
 import BusinessHero from "@/components/business/business-hero";
-import IdealClientsProfile from "./clients";
-import ValueProposition from "./value-proposition";
-import Services from "./services";
-import CaseStudies from "./case-studies";
-import TechStack from "./tech-stack";
-import Workflow from "./workflow";
-import SocialProofs from "./social-proofs";
+import IdealClientsProfile from "@/components/business/clients";
+import ValueProposition from "@/components/business/value-proposition";
+import Services from "@/components/business/services";
+import CaseStudies from "@/components/business/case-studies";
+import TechStack from "@/components/business/tech-stack";
+import Workflow from "@/components/business/workflow";
+import SocialProofs from "@/components/business/social-proofs";
 
 export default function BusinessPage() {
   const meetingLink = process.env.NEXT_PUBLIC_CALENDLY_URL!;
@@ -28,13 +28,16 @@ export default function BusinessPage() {
       <Workflow />
       <SocialProofs />
 
-      <section id="contact" className="py-20 px-4 md:px-6 bg-gradient-to-br from-[#392467] to-[#5a3a8f] text-white" style={{ scrollMarginTop: '80px' }}>
+      <section
+        id="contact"
+        className="py-20 px-4 md:px-6 bg-gradient-to-br from-[#392467] to-[#5a3a8f] text-white"
+        style={{ scrollMarginTop: "80px" }}
+      >
         <div className="container mx-auto max-w-4xl text-center pt-6 rounded-sm">
-          <h2 className="text-3xl md:text-5xl font-bold mb-6">
-            Ready to Build Something Great?
-          </h2>
+          <h2 className="text-3xl md:text-5xl font-bold mb-6">Ready to Build Something Great?</h2>
           <p className="text-lg md:text-xl mb-8 opacity-90 max-w-2xl mx-auto leading-relaxed">
-            Every project is value-based and scoped according to business goals. Book a free strategy call to get clarity on your technical roadmap and next steps.
+            Every project is value-based and scoped according to business goals. Book a free
+            strategy call to get clarity on your technical roadmap and next steps.
           </p>
 
           <div className="flex flex-col sm:flex-row gap-4 justify-center items-center">
@@ -55,14 +58,14 @@ export default function BusinessPage() {
               className="border-2 border-white text-whites hover:bg-white/10 text-base px-8 py-6 rounded-lg"
               asChild
             >
-              <Link href="mailto:your@email.com">
-                Or Email Me Directly
-              </Link>
+              <Link href="mailto:your@email.com">Or Email Me Directly</Link>
             </Button>
           </div>
 
           <div className="mt-12 pt-12 border-t border-white/20">
-            <p className="text-sm md:text-base opacity-75 mb-4">Trusted technical partner for startups and scale-ups worldwide</p>
+            <p className="text-sm md:text-base opacity-75 mb-4">
+              Trusted technical partner for startups and scale-ups worldwide
+            </p>
             <div className="flex justify-center space-x-8 text-sm md:text-base">
               <span className="flex items-center space-x-2">
                 <CheckCircle2 className="hidden md:block h-4 w-4" />
@@ -87,13 +90,28 @@ export default function BusinessPage() {
               © {new Date().getFullYear()} Oladipupo Ishola. All rights reserved.
             </p>
             <div className="flex space-x-6">
-              <Link href="https://github.com/olaishola05" target="_blank" rel="noopener noreferrer" className="text-softTextColor hover:text-[var(--primary)] transition-colors">
+              <Link
+                href="https://github.com/olaishola05"
+                target="_blank"
+                rel="noopener noreferrer"
+                className="text-softTextColor hover:text-primary transition-colors"
+              >
                 GitHub
               </Link>
-              <Link href="https://www.linkedin.com/in/olaishola05/" target="_blank" rel="noopener noreferrer" className="text-softTextColor hover:text-[var(--primary)] transition-colors">
+              <Link
+                href="https://www.linkedin.com/in/olaishola05/"
+                target="_blank"
+                rel="noopener noreferrer"
+                className="text-softTextColor hover:text-primary transition-colors"
+              >
                 LinkedIn
               </Link>
-              <Link href="https://twitter.com/olaishola05" target="_blank" rel="noopener noreferrer" className="text-softTextColor hover:text-[var(--primary)] transition-colors">
+              <Link
+                href="https://twitter.com/olaishola05"
+                target="_blank"
+                rel="noopener noreferrer"
+                className="text-softTextColor hover:text-primary transition-colors"
+              >
                 Twitter
               </Link>
             </div>
