+++ src/components/business/value-proposition.tsx
